db-agent/pyproject.toml
[tool.poetry]
name = "db-agent"
version = "0.1.0"
description = ""
authors = ["Muhammad Junaid <muhammadjunaid7015140@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.22"
fastapi = "^0.115.6"
langchain-community = "^0.3.14"
langgraph = "^0.2.60"
langchain-google-genai = "^2.0.7"
langchain-core = "^0.3.29"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
mypy = "^1.14.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
